refactor(logic): move crypto functions into util package
